1.16.1 working

+++ _Server/Arduino/arduino/arduino.ino
@@ -1,7 +1,7 @@
 #define damage_pin 2
 
 unsigned long start = 0;
-int damage_time = 0;
+float damage_time = 0;
 bool damage_active = false;
 
 void setup() {
@@ -15,7 +15,7 @@ void setup() {
 
 void loop() {
   if (Serial.available() > 0) {
-    int number = Serial.parseInt();
+    float number = Serial.parseFloat();
     while (Serial.available() > 0 && Serial.read() != '\n');
     
     if (number > 0) {
@@ -26,7 +26,7 @@ void loop() {
     }
   }
   
-  if (damage_active && (millis() - start >= (unsigned long)damage_time * 1000)) {
+  if (damage_active && (millis() - start >= (unsigned long)(damage_time * 1000))) {
     digitalWrite(damage_pin, LOW);
     damage_active = false;
   }
